# src/generated/swagger.json
{
	"components": {
...
		"schemas": {
			"User": {
				"properties": {
					"isActive": {
						"type": "boolean"
					},
					"createAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"passwordHash": {
						"type": "string",
						"nullable": true
					},
					"email": {
						"type": "string",
						"nullable": true
					},
					"fullName": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"isActive",
					"createAt",
					"updatedAt",
					"passwordHash",
					"email",
					"fullName",
					"username",
					"id"
				],
				"type": "object",
				"description": "Model User"
			},
			"Photo": {
				"properties": {
					"isActive": {
						"type": "boolean"
					},
					"createAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"userId": {
						"type": "number",
						"format": "double"
					},
					"url": {
						"type": "string"
					},
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"isActive",
					"createAt",
					"updatedAt",
					"userId",
					"url",
					"id"
				],
				"type": "object",
				"description": "Model Photo"
			},
			"Group": {
				"properties": {
					"isActive": {
						"type": "boolean"
					},
					"createAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"description": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"isActive",
					"createAt",
					"updatedAt",
					"description",
					"name",
					"id"
				],
				"type": "object",
				"description": "Model Group"
			},
			"UserDetailDto": {
				"allOf": [
					{
						"$ref": "#/components/schemas/User"
					},
					{
						"properties": {
							"permissions": {
								"items": {
									"type": "string"
								},
								"type": "array"
							},
							"groups": {
								"items": {
									"$ref": "#/components/schemas/Group"
								},
								"type": "array"
							},
							"photos": {
								"items": {
									"$ref": "#/components/schemas/Photo"
								},
								"type": "array"
							}
						},
						"required": [
							"permissions",
							"groups",
							"photos"
						],
						"type": "object"
					}
				]
			},
			"UserDto": {
				"$ref": "#/components/schemas/User"
			},
...
			"PhotoDetailDto": {
				"allOf": [
					{
						"$ref": "#/components/schemas/Photo"
					},
					{
						"properties": {
							"user": {
								"properties": {
									"username": {
										"type": "string"
									},
									"id": {
										"type": "number",
										"format": "double"
									}
								},
								"required": [
									"username",
									"id"
								],
								"type": "object"
							}
						},
						"required": [
							"user"
						],
						"type": "object"
					}
				]
			},
			"UpdatePhotoDto": {
				"properties": {
					"url": {
						"type": "string"
					}
				},
				"required": [
					"url"
				],
				"type": "object"
			},
			"Paginate_Photo_": {
				"properties": {
					"rows": {
						"items": {
							"$ref": "#/components/schemas/Photo"
						},
						"type": "array"
					},
					"page": {
						"type": "number",
						"format": "double"
					},
					"perPage": {
						"type": "number",
						"format": "double"
					},
					"total": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"rows",
					"page",
					"perPage",
					"total"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PhotoListDto": {
				"$ref": "#/components/schemas/Paginate_Photo_"
			},
			"PhotoDto": {
				"$ref": "#/components/schemas/Photo"
			},
			"CreatePhotoDto": {
				"properties": {
					"url": {
						"type": "string"
					},
					"username": {
						"type": "string"
					}
				},
				"required": [
					"url",
					"username"
				],
				"type": "object"
			},
			"UpdateUserDto": {
				"properties": {
					"photos": {
						"items": {
							"properties": {
								"url": {
									"type": "string"
								}
							},
							"required": [
								"url"
							],
							"type": "object"
						},
						"type": "array"
					},
					"groups": {
						"items": {
							"properties": {
								"name": {
									"type": "string"
								}
							},
							"required": [
								"name"
							],
							"type": "object"
						},
						"type": "array"
					},
					"isActive": {
						"type": "boolean"
					},
					"password": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"fullName": {
						"type": "string"
					}
				},
				"type": "object"
			},
			"Paginate_User_": {
				"properties": {
					"rows": {
						"items": {
							"$ref": "#/components/schemas/User"
						},
						"type": "array"
					},
					"page": {
						"type": "number",
						"format": "double"
					},
					"perPage": {
						"type": "number",
						"format": "double"
					},
					"total": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"rows",
					"page",
					"perPage",
					"total"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserListDto": {
				"$ref": "#/components/schemas/Paginate_User_"
			},
			"CreateUserDto": {
				"properties": {
					"email": {
						"type": "string"
					},
					"fullName": {
						"type": "string"
					},
					"username": {
						"type": "string",
						"minLength": 1
					}
				},
				"required": [
					"fullName",
					"username"
				],
				"type": "object"
			}
		},
		"securitySchemes": {
			"basic": {
				"scheme": "basic",
				"type": "http"
			},
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"name": "Bearer Authentication"
			}
		}
	},
	"info": {
		"title": "rest-api",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/api/v1/users/photos/{id}": {
			"get": {
				"operationId": "GetPhotoDetail",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PhotoDetailDto"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdatePhoto",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PhotoDetailDto"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdatePhotoDto"
							}
						}
					}
				}
			}
		},
		"/api/v1/users/photos": {
			"get": {
				"operationId": "GetPhotoList",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PhotoListDto"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "perPage",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "username",
						"required": false,
						"schema": {
							"default": null,
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreatePhoto",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PhotoDto"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreatePhotoDto"
							}
						}
					}
				}
			}
		},
		"/api/v1/users/{id}": {
			"get": {
				"operationId": "GetUserDetail",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserDetailDto"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserDetailDto"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"userAdmin",
							"users:update"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateUserDto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserDto"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"userAdmin",
							"users:delete"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/v1/users": {
			"get": {
				"operationId": "GetUserList",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserListDto"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"userAdmin",
							"users:read"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "perPage",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "isActive",
						"required": false,
						"schema": {
							"default": true,
							"type": "boolean"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserDto"
								}
							}
						}
					}
				},
				"description": "Hola mundo",
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"userAdmin",
							"users:create"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateUserDto"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}